Move DbContextExtensions.cs to correct namespace Init-only IGuildObject.cs#GuildId
