+++ Plugins/UVOCBot.Plugins.Planetside/Commands/OutfitTrackingCommands.cs
@@ -10,6 +10,7 @@ using System.ComponentModel;
 using System.Text;
 using System.Threading.Tasks;
 using UVOCBot.Core;
+using UVOCBot.Core.Extensions;
 using UVOCBot.Core.Model;
 using UVOCBot.Discord.Core;
 using UVOCBot.Discord.Core.Abstractions.Services;
