+++ UVOCBot.Core/Extensions/DbContextExtensions.cs
@@ -3,29 +3,25 @@ using System.Threading;
 using System.Threading.Tasks;
 using UVOCBot.Core.Model;
 
-namespace UVOCBot.Core;
+namespace UVOCBot.Core.Extensions;
 
 public static class DbContextExtensions
 {
-    public static async ValueTask<TEntity> FindOrDefaultAsync<TEntity>(this DbContext context, ulong guildId, CancellationToken ct = default) where TEntity : class, IGuildObject, new()
+    public static async ValueTask<TEntity> FindOrDefaultAsync<TEntity>
+    (
+        this DbContext context,
+        ulong guildId,
+        CancellationToken ct = default
+    ) where TEntity : class, IGuildObject, new()
     {
         TEntity? value = await context.FindAsync<TEntity>(new object[] { guildId }, ct).ConfigureAwait(false);
 
         if (value is null)
         {
-            TEntity e = new()
-            {
-                GuildId = guildId
-            };
-
-            context.Add(e);
-            await context.SaveChangesAsync(ct).ConfigureAwait(false);
-
-            return e;
+            value = new TEntity { GuildId = guildId };
+            context.Add(value);
         }
-        else
-        {
-            return value;
-        }
+
+        return value;
     }
 }
