+++ UVOCBot.Core/Model/PlanetsideSettings.cs
@@ -7,7 +7,7 @@ public class PlanetsideSettings : IGuildObject
 {
     /// <inheritdoc />
     [Key]
-    public ulong GuildId { get; set; }
+    public ulong GuildId { get; init; }
 
     /// <summary>
     /// Gets or sets the default world to use
@@ -25,16 +25,10 @@ public class PlanetsideSettings : IGuildObject
     public ulong? BaseCaptureChannelId { get; set; }
 
     /// <summary>
-    /// Only use this constructor if you are setting the <see cref="GuildId"/> immediately after construction.
+    /// Initializes a new instance of the <see cref="PlanetsideSettings"/> class.
     /// </summary>
     public PlanetsideSettings()
-        : this(0)
     {
-    }
-
-    public PlanetsideSettings(ulong guildId)
-    {
-        GuildId = guildId;
         TrackedOutfits = new List<ulong>();
     }
 }
