+++ UVOCBot/Commands/AdminCommands.cs
@@ -10,6 +10,7 @@ using Remora.Results;
 using System.ComponentModel;
 using System.Threading.Tasks;
 using UVOCBot.Core;
+using UVOCBot.Core.Extensions;
 using UVOCBot.Core.Model;
 using UVOCBot.Discord.Core;
 using UVOCBot.Discord.Core.Abstractions.Services;
